+++ crest/Assets/Crest/Crest/Scripts/Interaction/BoatProbes.cs
@@ -6,6 +6,9 @@
 
 using System;
 using UnityEngine;
+#if ENABLE_INPUT_SYSTEM
+using UnityEngine.InputSystem;
+#endif
 using UnityEngine.Serialization;
 
 namespace Crest
@@ -140,11 +143,23 @@ namespace Crest
             var forcePosition = _rb.position;
 
             var forward = _engineBias;
-            if (_playerControlled) forward += Input.GetAxis("Vertical");
+            if (_playerControlled) forward +=
+#if ENABLE_INPUT_SYSTEM
+                (Keyboard.current.wKey.isPressed ? 1 : 0) + (Keyboard.current.sKey.isPressed ? -1 : 0);
+#else
+                Input.GetAxis("Vertical");
+#endif
             _rb.AddForceAtPosition(transform.forward * _enginePower * forward, forcePosition, ForceMode.Acceleration);
 
             var sideways = _turnBias;
-            if (_playerControlled) sideways += (Input.GetKey(KeyCode.A) ? -1f : 0f) + (Input.GetKey(KeyCode.D) ? 1f : 0f);
+            if (_playerControlled) sideways +=
+#if ENABLE_INPUT_SYSTEM
+                (Keyboard.current.aKey.isPressed ? -1f : 0f) +
+                (Keyboard.current.dKey.isPressed ? 1f : 0f);
+#else
+                (Input.GetKey(KeyCode.A) ? -1f : 0f) +
+                (Input.GetKey(KeyCode.D) ? 1f : 0f);
+#endif
             var rotVec = transform.up + _turningHeel * transform.forward;
             _rb.AddTorque(rotVec * _turnPower * sideways, ForceMode.Acceleration);
         }